New methods in the Configurator interface
- for the name of the header file about kinematics parameters

- for the names of the classes with the coordinate transforms;
deleted old functions with namespace names

The second point is a consequence of the modifications done in the
Transforms-DSL generators, where the transforms are no longer global
variables in a namespace, but public members of a class. The name of
such a class is the return value the new configuration methods added here.